Added floating point input validation

+++ Finance.cpp
@@ -28,7 +28,7 @@ string Finance::getItem()
 {
     return item;
 }
-double Finance::getAmount()
+float Finance::getAmount()
 {
     return amount;
 }
